fix(config): keep codename hash architecture-stable on 32-bit
The function hashed through u64 specifically to keep the index identical
across 32-bit and 64-bit, but the final cast was `(u64 as usize) % len` —
on 32-bit `as usize` truncates to 32 bits before the modulo, so the same
branch would pick a different word than on 64-bit, defeating the whole
point of the u64 indirection.

Take the modulo in u64 first, then narrow. No-op on 64-bit (the existing
`test_codename_outputs_are_stable` snapshots are unchanged), correct on
32-bit. Worktrunk doesn't ship 32-bit binaries today, so this is a
correctness fix for the comment more than for any user — but the comment
was making a load-bearing claim, so do the matching arithmetic.

Closes part of #2663.

+++ src/config/expansion.rs
@@ -496,7 +496,11 @@ fn codename_index(input: &str, position: usize, salt: usize, pool: &str, len: us
     let digest = hasher.finalize();
     let mut bytes = [0_u8; 8];
     bytes.copy_from_slice(&digest[..8]);
-    (u64::from_le_bytes(bytes) as usize) % len
+    // Take the modulo in u64 before narrowing — `as usize` on a 32-bit
+    // build truncates the upper 32 bits and would pick a different word
+    // than 64-bit for the same branch, defeating the whole point of
+    // hashing through a fixed-width type above.
+    (u64::from_le_bytes(bytes) % len as u64) as usize
 }
 
 fn codename(input: &str, words: usize) -> String {
